feat(telemetry): add one-time register snapshot and fix Grafana queries
- Add RegisterSnapshot sending on first FULL-level poll (coordinator)
- Fix RegisterSnapshot.registers type: dict[str, int] → dict[str, float]
- Log telemetry level at startup for easier debugging
- Fix Grafana dashboard SQL variable syntax for multi-value filters
- Apply ruff formatting to anonymizer, aggregator, sensors, and tests

+++ custom_components/hitachi_yutaki/telemetry/models.py
@@ -294,7 +294,7 @@ class RegisterSnapshot:
     time: datetime
     profile: str
     gateway_type: str
-    registers: dict[str, int]
+    registers: dict[str, float]
 
     def to_dict(self) -> dict:
         """Serialize to dict for JSON payload."""
